rev-mc: Add earlyclobber annotation to ARM fake dependency
In measure_one_block_access_time() the aarch64 implementation uses

    asm volatile(
        "add %0, %1, #0\n\t"
        "eor %0, %0, %2\n\t"
        "eor %0, %0, %2\n\t"
        : "=r"(safe_ptr)
        : "r"(p2), "r"(val)
        : "memory");

to create a fake dependency on val.  On a Raspberry Pi 4B running
Raspberry Pi OS, this is compiled by g++ 14.2.0-19 to

    342c:       91000020        add     x0, x1, #0x0
    3430:       ca000000        eor     x0, x0, x0
    3434:       ca000000        eor     x0, x0, x0

The compiler reuses the same register for the input %2 (val) and the
output %0 (safe_ptr), leading to the eor instruction clearing out
safe_ptr.  This triggers the following assertion:

    main: src/rev-mc.cpp:76: uint64_t measure_one_block_access_time(volatile void*, volatile void*): Assertion `safe_p2 == p2 && "safe_p2 should equal p2 (dependency transformation failed!)"' failed.
    Aborted

This patch adds an earlyclobber annotation (&) on safe_ptr
to avoid this register aliasing issue, leading to the intended code

    342c:       91000020        add     x0, x1, #0x0
    3430:       ca020000        eor     x0, x0, x2
    3434:       ca020000        eor     x0, x0, x2